# openapi/daraja-local.openapi.yaml
openapi: 3.1.0
info:
  title: Daraja Local API
  version: 0.1.0
  description: Unofficial local emulator for selected Safaricom Daraja API flows.
servers:
  - url: http://127.0.0.1:8080
...
paths:
...
  /simulator/transactions/{id}:
    get:
      tags: [Simulator]
      summary: Inspect a transaction
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          $ref: "#/components/responses/TransactionFound"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /simulator/stk/{id}/replay-callback:
    post:
      tags: [Simulator]
      summary: Replay an STK callback
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          $ref: "#/components/responses/TransactionFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /simulator/payments/{id}/replay-callback:
    post:
      tags: [Simulator]
      summary: Replay a business-payment callback
      description: Replays the current terminal callback; timeout callbacks are delivered to QueueTimeOutURL and other results to ResultURL.
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          $ref: "#/components/responses/TransactionFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    TransactionId:
      name: id
      in: path
      required: true
      description: Emulator tracking identifier.
      schema:
        type: string
        minLength: 1
...
  responses:
...
    TransactionFound:
      description: Transaction state, including recorded callback attempts.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Transaction"
    BadRequest:
      description: Request validation or callback configuration failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Transaction or C2B registration was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Transaction kind or state does not permit the operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    Transaction:
      type: object
      required: [id, kind, trackingId, status, amount, rawRequest, createdAt, updatedAt, callbackAttempts]
      properties:
        id: { type: string }
        kind:
          type: string
          enum: [STK, C2B, B2C, B2B, B2C_ACCOUNT_TOPUP]
        trackingId: { type: string }
        merchantRequestId: { type: string }
        checkoutRequestId: { type: string }
        conversationId: { type: string }
        originatorConversationId: { type: string }
        status:
          type: string
          enum: [PENDING, SUCCESS, FAILED, TIMEOUT]
        amount: { type: number }
        phoneNumber: { type: string }
        accountReference: { type: string }
        senderParty: { type: string }
        receiverParty: { type: string }
        commandId: { type: string }
        callbackUrl: { type: string, format: uri }
        resultUrl: { type: string, format: uri }
        queueTimeoutUrl: { type: string, format: uri }
        rawRequest: {}
        resultCode:
          oneOf:
            - type: integer
            - type: string
          description: Emulator result; production values are product- and version-sensitive.
          x-version-sensitive: true
        resultDesc: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        callbackAttempts:
          type: array
          items:
            $ref: "#/components/schemas/CallbackAttempt"
    CallbackAttempt:
      type: object
      required: [id, sentAt, ok, payload]
      properties:
        id: { type: string }
        sentAt: { type: string, format: date-time }
        statusCode: { type: integer }
        ok: { type: boolean }
        error: { type: string }
        payload: {}
        responseBody: {}
...